Added in menu sounds, 1st level music
Added in menu hover sounds and click sounds to the main menu and pause menu. I was tweaking with the volume of them a lot, so let me know if anyone thinks they should be louder or quieter.

Added the first level music in

I also added in the spirit hum script and put it on the bunny spirit for testing / if anybody wants to hear it. It is said to not start right now so it has no affect on the game right now, but you can check that box if you want to hear it.