Refactor for spirit hum, added a lot of control, features to it as well
Whole new script and work in FMOD to reformat how it is going to work.

It is currently set that all three spirits have different registers they hum in when in the spirit world. We may not want all three of them harmonizing all the time, or at all and we may want to lower the volume, but they vibing in there