Figured out a big ongoing thing, overhauled audio, added human drone sound, violin accomplishment sound, did a lot of other misc fixed
Figured out a big ongoing thing, overhauled audio as a result of it. Everything should work better because of it.
Added human drone sound, might be a tiny bit creepy at times, but I think overall helps a lot.
Added violin accomplishment sound, that plays when the player figures something out. Only in the key puzzle right now.
Made a music cue in the tutorial, that the drums and bass will only come in once you get through the gate
Did a lot of other misc fixes as well